Proposed paragraph style setting (Paragraph style vanishs after text/style update. Otherwise complete)
