+++ richtextfx-demos/src/main/java/org/fxmisc/richtext/demo/RichText.java
@@ -12,6 +12,8 @@ import java.util.function.Function;
 import javafx.application.Application;
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.BooleanBinding;
+import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.collections.FXCollections;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
@@ -27,6 +29,8 @@ import javafx.scene.text.Font;
 import javafx.stage.Stage;
 
 import org.fxmisc.richtext.InlineStyleTextArea;
+import org.fxmisc.richtext.Paragraph;
+import org.fxmisc.richtext.StyleSpan;
 import org.fxmisc.richtext.StyleSpans;
 import org.reactfx.SuspendableNo;
 
@@ -49,14 +53,16 @@ public class RichText extends Application {
         final Optional<Color> backgroundColor;
 
         public StyleInfo() {
-            bold = Optional.empty();
-            italic = Optional.empty();
-            underline = Optional.empty();
-            strikethrough = Optional.empty();
-            fontSize = Optional.empty();
-            fontFamily = Optional.empty();
-            textColor = Optional.empty();
-            backgroundColor = Optional.empty();
+            this(
+                Optional.empty(),
+                Optional.empty(),
+                Optional.empty(),
+                Optional.empty(),
+                Optional.empty(),
+                Optional.empty(),
+                Optional.empty(),
+                Optional.empty()
+            );
         }
 
         public StyleInfo(
@@ -201,6 +207,7 @@ public class RichText extends Application {
     }
 
     private final SuspendableNo updatingToolbar = new SuspendableNo();
+    private final BooleanProperty applyToParagraph = new SimpleBooleanProperty(false);
 
     @Override
     public void start(Stage primaryStage) {
@@ -222,6 +229,9 @@ public class RichText extends Application {
         familyCombo.getSelectionModel().select("Serif");
         ColorPicker textColorPicker = new ColorPicker(Color.BLACK);
         ColorPicker backgroundColorPicker = new ColorPicker(Color.WHITE);
+        CheckBox paragraphCheckBox = new CheckBox();
+        paragraphCheckBox.setText("Apply to paragraph");
+        this.applyToParagraph.bind(paragraphCheckBox.selectedProperty());
 
         sizeCombo.setOnAction(evt -> updateFontSize(sizeCombo.getValue()));
         familyCombo.setOnAction(evt -> updateFontFamily(familyCombo.getValue()));
@@ -337,7 +347,7 @@ public class RichText extends Application {
         });
 
         HBox panel = new HBox(3.0);
-        panel.getChildren().addAll(wrapToggle, undoBtn, redoBtn, cutBtn, copyBtn, pasteBtn, boldBtn, italicBtn, underlineBtn, strikeBtn, sizeCombo, familyCombo, textColorPicker, backgroundColorPicker);
+        panel.getChildren().addAll(wrapToggle, undoBtn, redoBtn, cutBtn, copyBtn, pasteBtn, boldBtn, italicBtn, underlineBtn, strikeBtn, sizeCombo, familyCombo, textColorPicker, backgroundColorPicker, paragraphCheckBox);
 
         VBox vbox = new VBox();
         VBox.setVgrow(area, Priority.ALWAYS);
@@ -380,21 +390,32 @@ public class RichText extends Application {
     }
 
     private void updateStyleInSelection(Function<StyleSpans<StyleInfo>, StyleInfo> mixinGetter) {
-        IndexRange selection = area.getSelection();
-        if(selection.getLength() != 0) {
-            StyleSpans<StyleInfo> styles = area.getStyleSpans(selection);
-            StyleInfo mixin = mixinGetter.apply(styles);
-            StyleSpans<StyleInfo> newStyles = styles.mapStyles(style -> style.updateWith(mixin));
-            area.setStyleSpans(selection.getStart(), newStyles);
+        if (applyToParagraph.get()) {
+            StyleInfo mixin = mixinGetter.apply(StyleSpans.singleton(new StyleSpan<>(StyleInfo.EMPTY, 0)));
+            Paragraph paragraph = area.getParagraph(area.getCurrentParagraph());
+            paragraph.setParagraphStyle(mixin);
+        } else {
+            IndexRange selection = area.getSelection();
+            if(selection.getLength() != 0) {
+                StyleSpans<StyleInfo> styles = area.getStyleSpans(selection);
+                StyleInfo mixin = mixinGetter.apply(styles);
+                StyleSpans<StyleInfo> newStyles = styles.mapStyles(style -> style.updateWith(mixin));
+                area.setStyleSpans(selection.getStart(), newStyles);
+            }
         }
     }
 
     private void updateStyleInSelection(StyleInfo mixin) {
-        IndexRange selection = area.getSelection();
-        if(selection.getLength() != 0) {
-            StyleSpans<StyleInfo> styles = area.getStyleSpans(selection);
-            StyleSpans<StyleInfo> newStyles = styles.mapStyles(style -> style.updateWith(mixin));
-            area.setStyleSpans(selection.getStart(), newStyles);
+        if (applyToParagraph.get()) {
+            Paragraph paragraph = area.getParagraph(area.getCurrentParagraph());
+            paragraph.setParagraphStyle(mixin);
+        } else {
+            IndexRange selection = area.getSelection();
+            if (selection.getLength() != 0) {
+                StyleSpans<StyleInfo> styles = area.getStyleSpans(selection);
+                StyleSpans<StyleInfo> newStyles = styles.mapStyles(style -> style.updateWith(mixin));
+                area.setStyleSpans(selection.getStart(), newStyles);
+            }
         }
     }
