+++ richtextfx/src/main/java/org/fxmisc/richtext/Paragraph.java
@@ -9,11 +9,14 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.value.ObservableObjectValue;
 import javafx.scene.control.IndexRange;
 
 import org.fxmisc.richtext.TwoDimensional.Position;
 
-public final class Paragraph<S> implements CharSequence {
+public final class Paragraph<S, PS> implements CharSequence {
 
     @SafeVarargs
     private static <T> List<T> list(T head, T... tail) {
@@ -30,35 +33,37 @@ public final class Paragraph<S> implements CharSequence {
     private final List<StyledText<S>> segments;
     private final Optional<LineTerminator> terminator;
     private final TwoLevelNavigator navigator;
+    private ObjectProperty<PS> paragraphStyle;
 
-    public Paragraph(String text, S style) {
-        this(new StyledText<S>(text, style));
+    public Paragraph(String text, S style, PS paragraphStyle) {
+        this(paragraphStyle, new StyledText<S>(text, style));
     }
 
     @SafeVarargs
-    public Paragraph(StyledText<S> text, StyledText<S>... texts) {
-        this(list(text, texts), Optional.empty());
+    public Paragraph(PS paragraphStyle, StyledText<S> text, StyledText<S>... texts) {
+        this(list(text, texts), paragraphStyle, Optional.empty());
     }
 
-    private Paragraph(String text, S style, LineTerminator terminator) {
-        this(text, style, Optional.of(terminator));
+    private Paragraph(String text, S style, PS paragraphStyle, LineTerminator terminator) {
+        this(text, style, paragraphStyle, Optional.of(terminator));
     }
 
-    private Paragraph(String text, S style, Optional<LineTerminator> terminator) {
-        this(new StyledText<>(text, style), terminator);
+    private Paragraph(String text, S style, PS paragraphStyle, Optional<LineTerminator> terminator) {
+        this(new StyledText<>(text, style), paragraphStyle, terminator);
     }
 
-    private Paragraph(StyledText<S> text, Optional<LineTerminator> terminator) {
-        this(Arrays.asList(text), terminator);
+    private Paragraph(StyledText<S> text, PS paragraphStyle, Optional<LineTerminator> terminator) {
+        this(Arrays.asList(text), paragraphStyle, terminator);
     }
 
-    private Paragraph(List<StyledText<S>> segments, LineTerminator terminator) {
-        this(segments, Optional.of(terminator));
+    private Paragraph(List<StyledText<S>> segments, PS paragraphStyle, LineTerminator terminator) {
+        this(segments, paragraphStyle, Optional.of(terminator));
     }
 
-    private Paragraph(List<StyledText<S>> segments, Optional<LineTerminator> terminator) {
+    private Paragraph(List<StyledText<S>> segments, PS paragraphStyle, Optional<LineTerminator> terminator) {
         assert !segments.isEmpty();
         this.segments = segments;
+        this.paragraphStyle = new SimpleObjectProperty<>(paragraphStyle);
         this.terminator = terminator;
         navigator = new TwoLevelNavigator(
                 () -> segments.size(),
@@ -94,12 +99,12 @@ public final class Paragraph<S> implements CharSequence {
         return terminator.isPresent();
     }
 
-    public Paragraph<S> terminate(LineTerminator terminator) {
+    public Paragraph<S, PS> terminate(LineTerminator terminator) {
         if(isTerminated()) {
             throw new IllegalStateException("Paragraph alread terminated.");
         }
 
-        return new Paragraph<>(segments, terminator);
+        return new Paragraph<>(segments, paragraphStyle.get(), terminator);
     }
 
     @Override
@@ -116,7 +121,7 @@ public final class Paragraph<S> implements CharSequence {
         return fullText().substring(from);
     }
 
-    public Paragraph<S> concat(Paragraph<S> p) {
+    public Paragraph<S, PS> concat(Paragraph<S, PS> p) {
         if(isTerminated()) {
             throw new IllegalStateException("Cannot append to a paragraph with line terminator.");
         }
@@ -129,7 +134,7 @@ public final class Paragraph<S> implements CharSequence {
             return this;
         }
 
-        StyledText<S> left = segments.get(segments.size()-1);
+        StyledText<S> left = segments.get(segments.size() - 1);
         StyledText<S> right = p.segments.get(0);
         if(Objects.equals(left.getStyle(), right.getStyle())) {
             StyledText<S> segment = left.concat(right);
@@ -137,16 +142,16 @@ public final class Paragraph<S> implements CharSequence {
             segs.addAll(segments.subList(0, segments.size()-1));
             segs.add(segment);
             segs.addAll(p.segments.subList(1, p.segments.size()));
-            return new Paragraph<S>(segs, p.terminator);
+            return new Paragraph<>(segs, paragraphStyle.get(), p.terminator);
         } else {
             List<StyledText<S>> segs = new ArrayList<>(segments.size() + p.segments.size());
             segs.addAll(segments);
             segs.addAll(p.segments);
-            return new Paragraph<S>(segs, p.terminator);
+            return new Paragraph<>(segs, paragraphStyle.get(), p.terminator);
         }
     }
 
-    public Paragraph<S> concat(CharSequence str) {
+    public Paragraph<S, PS> concat(CharSequence str) {
         if(isTerminated()) {
             throw new IllegalStateException("Cannot append to a paragraph with line terminator.");
         }
@@ -158,10 +163,10 @@ public final class Paragraph<S> implements CharSequence {
         List<StyledText<S>> segs = new ArrayList<>(segments);
         int lastIdx = segments.size() - 1;
         segs.set(lastIdx, segments.get(lastIdx).concat(str));
-        return new Paragraph<S>(segs, Optional.empty());
+        return new Paragraph<S, PS>(segs, paragraphStyle.get(), Optional.empty());
     }
 
-    public Paragraph<S> insert(int offset, CharSequence str) {
+    public Paragraph<S, PS> insert(int offset, CharSequence str) {
         if(offset < 0 || offset > length())
             throw new IndexOutOfBoundsException(String.valueOf(offset));
 
@@ -172,33 +177,33 @@ public final class Paragraph<S> implements CharSequence {
         StyledText<S> replacement = seg.spliced(segPos, segPos, str);
         List<StyledText<S>> segs = new ArrayList<>(segments);
         segs.set(segIdx, replacement);
-        return new Paragraph<S>(segs, terminator);
+        return new Paragraph<S, PS>(segs, paragraphStyle.get(), terminator);
     }
 
     @Override
-    public Paragraph<S> subSequence(int start, int end) {
+    public Paragraph<S, PS> subSequence(int start, int end) {
         return trim(end).subSequence(start);
     }
 
-    public Paragraph<S> trim(int length) {
+    public Paragraph<S, PS> trim(int length) {
         if(length >= fullLength()) {
             return this;
         } else if(length > length()) { // slicing the line terminator
             LineTerminator newTerminator = terminator.get().trim(length - length());
-            return new Paragraph<>(segments, newTerminator);
+            return new Paragraph<>(segments, paragraphStyle.get(), newTerminator);
         } else if(length == length()) { // cut off the line terminator
-            return new Paragraph<>(segments, Optional.empty());
+            return new Paragraph<>(segments, paragraphStyle.get(), Optional.empty());
         } else { // length < length()
             Position pos = navigator.offsetToPosition(length, Backward);
             int segIdx = pos.getMajor();
             List<StyledText<S>> segs = new ArrayList<>(segIdx + 1);
             segs.addAll(segments.subList(0, segIdx));
             segs.add(segments.get(segIdx).subSequence(0, pos.getMinor()));
-            return new Paragraph<S>(segs, Optional.empty());
+            return new Paragraph<S, PS>(segs, paragraphStyle.get(), Optional.empty());
         }
     }
 
-    public Paragraph<S> subSequence(int start) {
+    public Paragraph<S, PS> subSequence(int start) {
         if(start < 0) {
             throw new IllegalArgumentException("start must not be negative (was: " + start + ")");
         } else if(start == 0) {
@@ -209,49 +214,49 @@ public final class Paragraph<S> implements CharSequence {
             List<StyledText<S>> segs = new ArrayList<>(segments.size() - segIdx);
             segs.add(segments.get(segIdx).subSequence(pos.getMinor()));
             segs.addAll(segments.subList(segIdx + 1, segments.size()));
-            return new Paragraph<S>(segs, terminator);
+            return new Paragraph<S, PS>(segs, paragraphStyle.get(), terminator);
         } else if(start < fullLength()) {
             LineTerminator newTerminator = terminator.get().subSequence(start - length());
             S lastStyle = segments.get(segments.size() - 1).getStyle();
-            return new Paragraph<>("", lastStyle, newTerminator);
+            return new Paragraph<>("", lastStyle, paragraphStyle.get(), newTerminator);
         } else if(isTerminated()) {
             throw new IndexOutOfBoundsException(start + " not in [0, " + fullLength() + ")");
         } else if(start == fullLength()) { // == length()
             S lastStyle = segments.get(segments.size() - 1).getStyle();
-            return new Paragraph<>("", lastStyle);
+            return new Paragraph<>("", lastStyle, paragraphStyle.get());
         } else {
             throw new IndexOutOfBoundsException(start + " not in [0, " + length() + "]");
         }
     }
 
-    public Paragraph<S> delete(int start, int end) {
+    public Paragraph<S, PS> delete(int start, int end) {
         return trim(start).concat(subSequence(end));
     }
 
-    public Paragraph<S> restyle(S style) {
-        return new Paragraph<S>(toString(), style, terminator);
+    public Paragraph<S, PS> restyle(S style) {
+        return new Paragraph<>(toString(), style, paragraphStyle.get(), terminator);
     }
 
-    public Paragraph<S> restyle(int from, int to, S style) {
+    public Paragraph<S, PS> restyle(int from, int to, S style) {
         if(from >= length()) {
             return this;
         } else {
             to = Math.min(to, length());
-            Paragraph<S> left = subSequence(0, from);
-            Paragraph<S> middle = new Paragraph<S>(substring(from, to), style);
-            Paragraph<S> right = subSequence(to);
+            Paragraph<S, PS> left = subSequence(0, from);
+            Paragraph<S, PS> middle = new Paragraph<S, PS>(substring(from, to), style, paragraphStyle.get());
+            Paragraph<S, PS> right = subSequence(to);
             return left.concat(middle).concat(right);
         }
     }
 
-    public Paragraph<S> restyle(int from, StyleSpans<? extends S> styleSpans) {
+    public Paragraph<S, PS> restyle(int from, StyleSpans<? extends S> styleSpans) {
         int len = styleSpans.length();
         if(styleSpans.equals(getStyleSpans(from, from + len))) {
             return this;
         }
 
-        Paragraph<S> left = trim(from);
-        Paragraph<S> right = subSequence(from + len);
+        Paragraph<S, PS> left = trim(from);
+        Paragraph<S, PS> right = subSequence(from + len);
 
         String middleString = substring(from, from + len);
         List<StyledText<S>> middleSegs = new ArrayList<>(styleSpans.getSpanCount());
@@ -262,7 +267,7 @@ public final class Paragraph<S> implements CharSequence {
             middleSegs.add(new StyledText<>(text, span.getStyle()));
             offset = end;
         }
-        Paragraph<S> middle = new Paragraph<>(middleSegs, Optional.empty());
+        Paragraph<S, PS> middle = new Paragraph<>(middleSegs, paragraphStyle.get(), Optional.empty());
 
         return left.concat(middle).concat(right);
     }
@@ -382,5 +387,17 @@ public final class Paragraph<S> implements CharSequence {
             }
         }
         return fullText;
+    }
+
+    public PS getParagraphStyle() {
+        return paragraphStyle.get();
+    }
+
+    public ObservableObjectValue<PS> paragraphStyle() {
+        return paragraphStyle;
+    }
+
+    public void setParagraphStyle(PS paragraphStyle) {
+        this.paragraphStyle.set(paragraphStyle);
     }
 }
