+++ richtextfx/src/main/java/org/fxmisc/richtext/StyledTextArea.java
@@ -96,10 +96,10 @@ import org.reactfx.value.Var;
  */
 public class StyledTextArea<S, PS> extends Control
 implements
-        TextEditingArea<S>,
-        EditActions<S>,
-        ClipboardActions<S>,
-        NavigationActions<S>,
+        TextEditingArea<S, PS>,
+        EditActions<S, PS>,
+        ClipboardActions<S, PS>,
+        NavigationActions<S, PS>,
         UndoActions<S>,
         TwoDimensional {
 
@@ -265,7 +265,7 @@ implements
     @Override public final ObservableValue<String> textProperty() { return text; }
 
     // rich text
-    @Override public final StyledDocument<S> getDocument() { return content.snapshot(); };
+    @Override public final StyledDocument<S, PS> getDocument() { return content.snapshot(); };
 
     // length
     private final SuspendableVal<Integer> length;
@@ -305,8 +305,8 @@ implements
     @Override public final ObservableValue<Integer> caretColumnProperty() { return caretColumn; }
 
     // paragraphs
-    private final SuspendableList<Paragraph<S>> paragraphs;
-    @Override public ObservableList<Paragraph<S>> getParagraphs() {
+    private final SuspendableList<Paragraph<S, PS>> paragraphs;
+    @Override public ObservableList<Paragraph<S, PS>> getParagraphs() {
         return paragraphs;
     }
 
@@ -328,9 +328,9 @@ implements
     public final EventStream<PlainTextChange> plainTextChanges() { return plainTextChanges; }
 
     // rich text changes
-    private final SuspendableEventStream<RichTextChange<S>> richTextChanges;
+    private final SuspendableEventStream<RichTextChange<S, PS>> richTextChanges;
     @Override
-    public final EventStream<RichTextChange<S>> richChanges() { return richTextChanges; }
+    public final EventStream<RichTextChange<S, PS>> richChanges() { return richTextChanges; }
 
 
     /* ********************************************************************** *
@@ -409,7 +409,7 @@ implements
         this.applyStyle = applyStyle;
         this.applyParagraphStyle = applyParagraphStyle;
         this.preserveStyle = preserveStyle;
-        content = new EditableStyledDocument<>(initialStyle,initialParagraphStyle );
+        content = new EditableStyledDocument<>(initialStyle, initialParagraphStyle);
         paragraphs = LiveList.suspendable(content.getParagraphs());
 
         text = Val.suspendable(content.textProperty());
@@ -483,17 +483,17 @@ implements
         return paragraphs.get(paragraph).toString();
     }
 
-    public Paragraph<S> getParagraph(int index) {
+    public Paragraph<S, PS> getParagraph(int index) {
         return paragraphs.get(index);
     }
 
     @Override
-    public StyledDocument<S> subDocument(int start, int end) {
+    public StyledDocument<S, PS> subDocument(int start, int end) {
         return content.subSequence(start, end);
     }
 
     @Override
-    public StyledDocument<S> subDocument(int paragraphIndex) {
+    public StyledDocument<S, PS> subDocument(int paragraphIndex) {
         return content.subDocument(paragraphIndex);
     }
 
@@ -695,6 +695,15 @@ implements
     }
 
     /**
+     * Sets style for the whole paragraph.
+     */
+    public void setParagraphStyle(int paragraph, PS paragraphStyle) {
+        if (paragraph >= 0 && paragraph < paragraphs.size()) {
+            paragraphs.get(paragraph).setParagraphStyle(paragraphStyle);
+        }
+    }
+
+    /**
      * Resets the style of the given range to the initial style.
      */
     public void clearStyle(int from, int to) {
@@ -716,6 +725,13 @@ implements
         setStyle(paragraph, from, to, initialStyle);
     }
 
+    /**
+     * Resets the style of the given paragraph to the initial style.
+     */
+    public void clearParagraphStyle(int paragraph) {
+        setParagraphStyle(paragraph, initialParagraphStyle);
+    }
+
     @Override
     public void replaceText(int start, int end, String text) {
         try(Guard g = omniSuspendable.suspend()) {
@@ -730,7 +746,7 @@ implements
     }
 
     @Override
-    public void replace(int start, int end, StyledDocument<S> replacement) {
+    public void replace(int start, int end, StyledDocument<S, PS> replacement) {
         try(Guard g = omniSuspendable.suspend()) {
             start = clamp(0, start, getLength());
             end = clamp(0, end, getLength());
@@ -802,9 +818,9 @@ implements
     }
 
     private UndoManager createRichUndoManager(UndoManagerFactory factory) {
-        Consumer<RichTextChange<S>> apply = change -> replace(change.getPosition(), change.getPosition() + change.getRemoved().length(), change.getInserted());
-        Consumer<RichTextChange<S>> undo = change -> replace(change.getPosition(), change.getPosition() + change.getInserted().length(), change.getRemoved());
-        BiFunction<RichTextChange<S>, RichTextChange<S>, Optional<RichTextChange<S>>> merge = (change1, change2) -> change1.mergeWith(change2);
+        Consumer<RichTextChange<S, PS>> apply = change -> replace(change.getPosition(), change.getPosition() + change.getRemoved().length(), change.getInserted());
+        Consumer<RichTextChange<S, PS>> undo = change -> replace(change.getPosition(), change.getPosition() + change.getInserted().length(), change.getRemoved());
+        BiFunction<RichTextChange<S, PS>, RichTextChange<S, PS>, Optional<RichTextChange<S, PS>>> merge = (change1, change2) -> change1.mergeWith(change2);
         return factory.create(richChanges(), apply, undo, merge);
     }
